Update README.md
Added links to documentation for direct referencing of the URLs. Provided context for articles that assist with onboarding, with each article specifically linked to the corresponding blog entry. Reorganized the ‘Getting Started’ section to highlight key actions, starting with the ability to connect with developers and access support, followed by the quick start guides.